Fixed demux buffer full

+++ ue/src/mac/pdu.cc
@@ -531,7 +531,7 @@ int sch_subh::set_sdu(uint32_t lcid_, uint32_t requested_bytes, rlc_interface_ma
 
     ((sch_pdu*)parent)->add_sdu(nof_bytes);
     ((sch_pdu*)parent)->update_space_sdu(nof_bytes);
-    return (int) nof_bytes; 
+    return nof_bytes; 
   } else {
     return -1; 
   }
